+++ src/components/CombinationIngredientList.tsx
@@ -1,6 +1,6 @@
 import styled from '@emotion/styled';
 import { changeRem } from '../styles/mixin';
-import {} from './SandwitchInfoCard';
+import {} from './UI/Cards/SandwitchInfoCard';
 
 interface 인터페이스_재료 {
   id: string;
