+++ src/components/CustomCombination/MyCombinationStep.tsx
@@ -1,7 +1,7 @@
 import styled from '@emotion/styled';
-import MyCombinationStepBadge from './MyCombinationStepBadge';
-import { changeRem } from '../styles/mixin';
-import mediaQuery from '../styles/media-queries';
+import { changeRem } from '@styles/mixin';
+import mediaQuery from '@styles/media-queries';
+import MyCombinationStepBadge from '@components/CustomCombination/MyCombinationStepBadge';
 
 interface 인터페이스_조합진행도_속성 {
   currentStep: number;
