--- src/components/MainNavigation.jsx
@@ -1,26 +1,0 @@
-import { NavLink } from 'react-router-dom';
-
-import classes from './MainNavigation.module.css';
-
-function MainNavigation() {
-  return (
-    <header className={classes.header}>
-      <nav>
-        <ul>
-          <li>
-            <NavLink to="/" className={({ isActive }) => (isActive ? classes.active : undefined)} end>
-              Home
-            </NavLink>
-          </li>
-          <li>
-            <NavLink to="/test" className={({ isActive }) => (isActive ? classes.active : undefined)} end>
-              test Page
-            </NavLink>
-          </li>
-        </ul>
-      </nav>
-    </header>
-  );
-}
-
-export default MainNavigation;
