--- src/components/MainNavigation.module.css
@@ -1,23 +1,0 @@
-.header {
-  padding: 2rem;
-  background-color: #924901;
-}
-
-.header ul {
-  display: flex;
-  gap: 2rem;
-  justify-content: center;
-  list-style: none;
-}
-
-.header a {
-  color: #e5e5e5;
-  font-size: 1.15rem;
-  text-decoration: none;
-}
-
-.header a:hover,
-.header a:active,
-.header a.active {
-  color: #fcb66b;
-}
