+++ tsconfig.json
@@ -108,7 +108,9 @@
       "@components/*": ["src/components/*"],
       "@constants/*": ["src/constants/*"],
       "@layouts/*": ["src/layouts/*"],
-      "@pages/*": ["src/pages/*"]
+      "@pages/*": ["src/pages/*"],
+      "@styles/*": ["src/styles/*"],
+      "@types/*": ["src/types/*"]
       // "@hooks/*": ["src/hooks/*"],
       // "@typings/*": ["typings/*"]
     }
